Fix imports after files were moved
