+++ src/components/Main.ui.js
@@ -78,7 +78,10 @@ import * as Constants from '../constants/indexConstants'
 import { scale } from '../lib/scaling.js'
 import { setIntlLocale } from '../locales/intl'
 import s, { selectLocale } from '../locales/strings.js'
+import EdgeAccountCallbackManager from '../modules/Core/Account/EdgeAccountCallbackManager.js'
 import * as CONTEXT_API from '../modules/Core/Context/api'
+import EdgeContextCallbackManager from '../modules/Core/Context/EdgeContextCallbackManager.js'
+import EdgeWalletsCallbackManager from '../modules/Core/Wallets/EdgeWalletsCallbackManager.js'
 import PermissionsManager, { type Permission, PermissionStrings } from '../modules/PermissionsManager.js'
 import AutoLogout from '../modules/UI/components/AutoLogout/AutoLogoutConnector'
 import { ContactsLoaderConnecter as ContactsLoader } from '../modules/UI/components/ContactsLoader/indexContactsLoader.js'
@@ -103,9 +106,6 @@ import { CreateWalletName } from './scenes/CreateWalletNameScene.js'
 import { CryptoExchangeQuoteProcessingScreenComponent } from './scenes/CryptoExchangeQuoteProcessingScene.js'
 import { OnBoardingComponent } from './scenes/OnBoardingScene.js'
 import { TermsOfServiceComponent } from './scenes/TermsOfServiceScene.js'
-import EdgeAccountCallbackManager from './Core/Account/EdgeAccountCallbackManager.js'
-import EdgeContextCallbackManager from './Core/Context/EdgeContextCallbackManager.js'
-import EdgeWalletsCallbackManager from './Core/Wallets/EdgeWalletsCallbackManager.js'
 
 const pluginFactories: Array<EdgeCorePluginFactory> = [
   // Exchanges:
